+++ applications/plugins/SofaCUDA/sofa/gpu/cuda/CudaBarycentricMapping-3f1.cpp
@@ -50,7 +50,7 @@ using namespace sofa::gpu::cuda;
 template<>
 void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::apply( Out::VecCoord& out, const In::VecCoord& in )
 {
-    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
     out.fastResize(map.size());
     RegularGridMapperCuda3f1_apply(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 }
@@ -58,7 +58,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::appl
 template<>
 void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::applyJ( Out::VecDeriv& out, const In::VecDeriv& in )
 {
-    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
     out.fastResize(map.size());
     RegularGridMapperCuda3f1_applyJ(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 }
@@ -68,7 +68,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::appl
 {
     calcMapT();
     if (map.size() == 0) return;
-    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
     unsigned int insize = out.size();
 
     RegularGridMapperCuda3f1_applyJT(insize, maxNOut, gridsize, mapT.deviceRead(), out.deviceWrite(), in.deviceRead());
@@ -90,7 +90,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::draw
 //template<>
 //void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3fTypes>::apply( Out::VecCoord& out, const In::VecCoord& in )
 //{
-//    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+//    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
 //    out.fastResize(map.size());
 //    RegularGridMapperCuda3f1_3f_apply(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 //}
@@ -98,7 +98,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::draw
 //template<>
 //void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3fTypes>::applyJ( Out::VecDeriv& out, const In::VecDeriv& in )
 //{
-//    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+//    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
 //    out.fastResize(map.size());
 //    RegularGridMapperCuda3f1_3f_applyJ(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 //}
@@ -108,7 +108,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::draw
 //{
 //    calcMapT();
 //    if (map.size() == 0) return;
-//    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+//    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
 //    unsigned int insize = out.size();
 
 //    RegularGridMapperCuda3f1_3f_applyJT(insize, maxNOut, gridsize, mapT.deviceRead(), out.deviceWrite(), in.deviceRead());
@@ -129,7 +129,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::draw
 //template<>
 //void BarycentricMapperRegularGridTopology<CudaVec3fTypes,CudaVec3f1Types>::apply( Out::VecCoord& out, const In::VecCoord& in )
 //{
-//    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+//    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
 //    out.fastResize(map.size());
 //    RegularGridMapperCuda3f_3f1_apply(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 //}
@@ -137,7 +137,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::draw
 //template<>
 //void BarycentricMapperRegularGridTopology<CudaVec3fTypes,CudaVec3f1Types>::applyJ( Out::VecDeriv& out, const In::VecDeriv& in )
 //{
-//    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+//    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
 //    out.fastResize(map.size());
 //    RegularGridMapperCuda3f_3f1_applyJ(map.size(), gridsize, map.deviceRead(), out.deviceWrite(), in.deviceRead());
 //}
@@ -147,7 +147,7 @@ void BarycentricMapperRegularGridTopology<CudaVec3f1Types,CudaVec3f1Types>::draw
 //{
 //    calcMapT();
 //    if (map.size() == 0) return;
-//    unsigned int gridsize[3] = { topology->getNx(), topology->getNy(), topology->getNz() };
+//    unsigned int gridsize[3] = { (unsigned int)topology->getNx(), (unsigned int)topology->getNy(), (unsigned int)topology->getNz() };
 //    unsigned int insize = out.size();
 
 //    RegularGridMapperCuda3f_3f1_applyJT(insize, maxNOut, gridsize, mapT.deviceRead(), out.deviceWrite(), in.deviceRead());
